requirements.txt
langchain
langchain-openai
langchainhub
langchain_community
chromadb
python-dotenv